Move some functions from plugins.utils to fields
