+++ rows/fields.py
@@ -26,17 +26,20 @@ from base64 import b64decode, b64encode
 from collections import OrderedDict
 from decimal import Decimal, InvalidOperation
 from textwrap import dedent
+from unicodedata import normalize
 
 import six
+
 
 # Order matters here
 __all__ = ['BoolField', 'IntegerField', 'FloatField', 'DatetimeField',
            'DateField', 'DecimalField', 'PercentField', 'JSONField',
            'EmailField', 'TextField', 'BinaryField', 'Field']
-REGEXP_ONLY_NUMBERS = re.compile('[^0-9\-]')
-SHOULD_NOT_USE_LOCALE = True  # This variable is changed by rows.locale_manager
 NULL = ('-', 'null', 'none', 'nil', 'n/a', 'na')
 NULL_BYTES = (b'-', b'null', b'none', b'nil', b'n/a', b'na')
+REGEXP_ONLY_NUMBERS = re.compile('[^0-9\-]')
+SHOULD_NOT_USE_LOCALE = True  # This variable is changed by rows.locale_manager
+SLUG_CHARS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_"
 
 
 def value_error(value, cls):
@@ -474,6 +477,75 @@ def unique_values(values):
     for value in values:
         if not is_null(value) and value not in result:
             result.append(value)
+    return result
+
+
+def slug(text, separator="_", permitted_chars=SLUG_CHARS, replace_with_separator=" -_"):
+    """Generate a slug for the `text`.
+
+    >>> slug(' ÁLVARO  justen% ')
+    'alvaro_justen'
+    >>> slug(' ÁLVARO  justen% ', separator='-')
+    'alvaro-justen'
+    """
+
+    text = six.text_type(text or "")
+
+    # Strip non-ASCII characters
+    # Example: u' ÁLVARO  justen% ' -> ' ALVARO  justen% '
+    text = normalize("NFKD", text.strip()).encode("ascii", "ignore").decode("ascii")
+
+    # Replace spaces and other chars with separator
+    # Example: u' ALVARO  justen% ' -> u'_ALVARO__justen%_'
+    for char in replace_with_separator:
+        text = text.replace(char, separator)
+
+    # Remove non-permitted characters and put everything to lowercase
+    # Example: u'_ALVARO__justen%_' -> u'_alvaro__justen_'
+    text = "".join(char for char in text if char in permitted_chars).lower()
+
+    # Remove double occurrencies of separator
+    # Example: u'_alvaro__justen_' -> u'_alvaro_justen_'
+    double_separator = separator + separator
+    while double_separator in text:
+        text = text.replace(double_separator, separator)
+
+    # Strip separators
+    # Example: u'_alvaro_justen_' -> u'alvaro_justen'
+    return text.strip(separator)
+
+
+def make_unique_name(name, existing_names, name_format="{name}_{index}", start=2):
+    """Return a unique name based on `name_format` and `name`."""
+    index = start
+    new_name = name
+    while new_name in existing_names:
+        new_name = name_format.format(name=name, index=index)
+        index += 1
+
+    return new_name
+
+
+def make_header(field_names, permit_not=False):
+    """Return unique and slugged field names."""
+    slug_chars = SLUG_CHARS if not permit_not else SLUG_CHARS + "^"
+
+    header = [
+        slug(field_name, permitted_chars=slug_chars) for field_name in field_names
+    ]
+    result = []
+    for index, field_name in enumerate(header):
+        if not field_name:
+            field_name = "field_{}".format(index)
+        elif field_name[0].isdigit():
+            field_name = "field_{}".format(field_name)
+
+        if field_name in result:
+            field_name = make_unique_name(
+                name=field_name, existing_names=result, start=2
+            )
+        result.append(field_name)
+
     return result
